move to apollo-client-config-data
