+++ apollo-client-config-data/src/main/java/com/ctrip/framework/apollo/config/data/importer/ApolloConfigDataLoaderInitializer.java
@@ -18,11 +18,11 @@ package com.ctrip.framework.apollo.config.data.importer;
 
 import com.ctrip.framework.apollo.config.data.extension.initialize.ApolloClientExtensionInitializeFactory;
 import com.ctrip.framework.apollo.config.data.injector.ApolloConfigDataInjectorCustomizer;
+import com.ctrip.framework.apollo.config.data.internals.PureApolloConfigFactory;
 import com.ctrip.framework.apollo.config.data.system.ApolloClientSystemPropertyInitializer;
 import com.ctrip.framework.apollo.config.data.util.Slf4jLogMessageFormatter;
 import com.ctrip.framework.apollo.core.utils.DeferredLogger;
-import com.ctrip.framework.apollo.spi.PureApolloRepositoryConfigFactory;
-import com.ctrip.framework.apollo.spi.RepositoryConfigFactory;
+import com.ctrip.framework.apollo.spi.ConfigFactory;
 import com.ctrip.framework.apollo.spring.config.PropertySourcesConstants;
 import java.util.Arrays;
 import java.util.Collections;
@@ -103,8 +103,8 @@ class ApolloConfigDataLoaderInitializer {
     new ApolloClientExtensionInitializeFactory(this.log,
         this.bootstrapContext).initializeExtension(this.binder, this.bindHandler);
     DeferredLogger.enable();
-    ApolloConfigDataInjectorCustomizer.register(RepositoryConfigFactory.class,
-        PureApolloRepositoryConfigFactory::new);
+    ApolloConfigDataInjectorCustomizer.register(ConfigFactory.class,
+        PureApolloConfigFactory::new);
   }
 
   private boolean forceDisableApolloBootstrap() {
